added what will be 2 more images for stories. instead of tracking the data of images i will just track the data of  videos to limit them by data that way. data is now being counted by doubles so i can show them in mega bytes such as 4.3 megs left for videos usesage. they are only allowed x amount fo memorials and stories. atm they can only have 3 memorials. they can hav unlimited stories atm but that will bet capped next. removed all the complex system of tracking image data size. now you will only be judge on how many images not how much data they take up.

+++ GraveYardApp/Memorial/UserProfile.swift
@@ -21,7 +21,7 @@ protocol DocumentSerializableUser {
 struct UserProfile {
     var currentUserAuthId: String
     var premiumStatus: Bool
-    var dataCount: Int
+    var dataCount: Double
     var memorialCount: Int
     
     var dictionary: [String: Any] {
@@ -38,7 +38,7 @@ extension UserProfile: DocumentSerializableUser {
     init?(dictionary: [String: Any]) {
         guard let currentUserAuthId = dictionary["currentUserAuthId"] as? String,
             let premiumStatus = dictionary["premiumStatus"] as? Bool,
-            let dataCount = dictionary["dataCount"] as? Int,
+            let dataCount = dictionary["dataCount"] as? Double,
             let memorialCount = dictionary["memorialCount"] as? Int else { return nil }
         self.init(currentUserAuthId: currentUserAuthId, premiumStatus: premiumStatus, dataCount: dataCount, memorialCount: memorialCount)
     }
